Added editText, button, and string values.
In this commit, I altered the dp margins for the app layout from 8 to 16 to accomodate curved screens. I created a default button and editText view, which have both been anchored to the margins while also creating and attaching a string value each entity. I have also connected both the button and edit text as far as functionality goes.