Added Display Child.
Added: Display child linked to the boot screen. This was achieved setup in the andriod manifest file.
Added: Send button functionality to read in inputted string in EditText and output that text to the new child.
Added: TextView in the child that is created by the use of Send button in parent.