VendMe 2.0a
Added: Secondary Fragmented Google Maps view on the second screen.

Added: Button to hold the number of vending machines found in the area to be selected by the user.

Added: Precautionary measure making the "Search This Area" button invisible or disabled when the user deny's to give location permission. This is to stop the user from proceeding through the app and causing a crash due to a lack of location data.