Vendme 2.0b
Added: Experimenting with WebViews to launch ISU's website when the markr ISU is clicked. (Trying to make this stay within the app and not launch an extrenal browser.)

Added: New activity and class to handle the WebView.

Removed: DisplayMessageActivity class and relating files.

Added: MachineSelection class as well as its activity.

Added: onMarkerClick method to listen for when a marker get clicked.